# internal/httpserver/swaggerdocs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
...
        "title": "GoDNS API",
        "contact": {
            "name": "GoDNS Support",
            "url": "https://github.com/rogerwesterbo/godns",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/rogerwesterbo/godns/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "localhost:14000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/export": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Export all DNS zones in a specified format (coredns, powerdns, bind, zonefile)",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Export"
                ],
                "summary": "Export all zones",
                "parameters": [
                    {
                        "type": "string",
                        "default": "bind",
                        "description": "Export format: coredns, powerdns, bind, or zonefile",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exported zone configuration",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/export/{zone}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "Export a specific DNS zone in a specified format",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Export"
                ],
                "summary": "Export a specific zone",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Zone name (e.g., example.lan)",
                        "name": "zone",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "bind",
                        "description": "Export format: coredns, powerdns, bind, or zonefile",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exported zone configuration",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid format",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Zone not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "http://localhost:14101/realms/godns/protocol/openid-connect/token"
        }
    },
...
}